requirements.txt
#LLM
openai>=1.0.0
langchain>=0.1.0
langgraph>=0.1.0

#stock market data
yfinance>=0.2.36

#Technical Indicators
pandas-ta>=0.3.14b

#Data Processing
pandas>=2.0.0
numpy>=1.24.0

#News / web search
tavily-python>=0.3.0

#validation (python's zod)
pydantic>=2.0.0

#environment variables
python-dotenv>=1.0.0

#Async HTTP
aiohttp>=3.9.0

#logging & observability
structlog>=23.0.0

#API Layer 
fastapi>=0.104.0
uvicorn>=0.24.0 

#Database
sqlalchemy>=2.0.0

#HTTP client (NewsAPI, CoinGecko, Telegram)
requests>=2.31.0

#Scheduling
schedule>=1.2.0

alpaca-py>=0.8.0